requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
autopep8==2.0.2
base58==2.1.1
bitarray==2.7.3
blinker==1.8.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cytoolz==0.12.1
distlib==0.3.6
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.9.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-MySQLdb==2.0.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hexbytes==0.3.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
ipfshttpclient==0.8.0a2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.5
multiaddr==0.0.9
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.2.4
netaddr==0.8.0
parsimonious==0.8.1
pipenv==2023.2.18
platformdirs==3.1.1
protobuf==3.20.3
py-mon==1.1.0
pycodestyle==2.10.0
pycoingecko==3.1.0
pycryptodome==3.17
PyJWT==2.8.0
pyperclip==1.8.2
pyrsistent==0.19.3
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.1
python-telegram-bot-calendar==1.0.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
random2==1.0.1
requests==2.28.2
rfc3986==1.5.0
rlp==2.0.1
six==1.16.0
sniffio==1.3.0
telebot==0.0.5
telegram==0.0.1
threaded==4.1.0
tomli==2.0.1
toolz==0.12.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
varint==1.0.2
virtualenv==20.21.0
virtualenv-clone==0.5.7
watchdog==2.3.1
web3==6.0.0b1
websockets==10.4
Werkzeug==3.0.3
yarl==1.8.2
